did sprite shape reading and constructing

+++ trunk/ascii_game_engine/render/agerenderer.h
@@ -61,7 +61,7 @@ typedef struct {
 	Frame* frames;
 	s32 frameCount;
 	s32 currentFrame;
-	s32 frameRate;
+	f32 frameRate;
 	MessageMap messageMap;
 	ControlProc control;
 } Sprite;
